Refactored express server methods to reduce complexity
