+++ lib/mocker/express/server.js
@@ -8,8 +8,7 @@ const cors = require('cors');
 const cookieParser = require('cookie-parser');
 const logger = require('lllog')();
 const colors = require('colors');
-const parsePreferHeader = require('parse-prefer-header');
-const memoize = require('micro-memoize');
+const handleRequest = require('./request-handler');
 
 const openApiMockSymbol = Symbol('openApiMock');
 
@@ -66,68 +65,33 @@ class Server {
 
 		this._loadBasePaths();
 
-		this.paths.map(path => {
-
-			logger.debug(`Processing schema path ${path.httpMethod.toUpperCase()} ${path.uri}`);
-
-			const expressHttpMethod = path.httpMethod.toLowerCase();
-
-			const uris = this._normalizeExpressPath(path.uri);
-
-			// Create a function that is memoized using the URL, query, the Prefer header and the body.
-			// eslint-disable-next-line no-unused-vars
-			const getResponse = (url, query, preferHeader, body) => {
-				const { example: preferredExampleName, statuscode: preferredStatusCode } = parsePreferHeader(preferHeader) || {};
-
-				if(preferredStatusCode)
-					logger.debug(`Searching requested response with status code ${preferredStatusCode}`);
-				else
-					logger.debug('Searching first response');
-				return path.getResponse(preferredStatusCode, preferredExampleName);
-			};
-
-			const getResponseMemo = memoize(getResponse, {
-				maxSize: 10
-			});
-
-			app[expressHttpMethod](uris, (req, res) => {
-
-				this._checkContentType(req);
-
-				const {
-					query,
-					params,
-					headers,
-					cookies,
-					body: requestBody
-				} = req;
-
-				const failedValidations = path.validateRequestParameters({
-					query,
-					path: params,
-					headers,
-					cookies,
-					requestBody
-				});
-
-				if(failedValidations.length)
-					return this.sendResponse(req, res, { errors: failedValidations }, 400);
-
-				const preferHeader = req.header('prefer') || '';
-
-				const { statusCode, headers: responseHeaders, body } =
-					getResponseMemo(req.path, JSON.stringify(req.query), preferHeader, JSON.stringify(requestBody));
-
-				return this.sendResponse(req, res, body, statusCode, responseHeaders);
-			});
-
-			return uris.map(uri => {
-				return logger.info(`Handling route ${path.httpMethod.toUpperCase()} ${uri}`);
-			});
-		});
+		this.setRoutes(app);
 
 		app.all('*', this._notFoundHandler.bind(this));
 
+		return this.startServer(app);
+	}
+
+	setRoutes(app) {
+		this.paths.map(path => this.setRoute(app, path));
+	}
+
+	setRoute(app, path) {
+
+		logger.debug(`Processing schema path ${path.httpMethod.toUpperCase()} ${path.uri}`);
+
+		const expressHttpMethod = path.httpMethod.toLowerCase();
+
+		const uris = this._normalizeExpressPath(path.uri);
+
+		app[expressHttpMethod](uris, handleRequest(path));
+
+		return uris.map(uri => {
+			return logger.info(`Handling route ${path.httpMethod.toUpperCase()} ${uri}`);
+		});
+	}
+
+	startServer(app) {
 		return new Promise((resolve, reject) => {
 			this.server = app.listen(this.port);
